Tidy up connector components
