+++ packages/lib-classifier/src/components/Classifier/components/TaskArea/components/DisabledTaskPopup/index.js
@@ -1,1 +1,1 @@
-export { default } from './DisabledTaskPopup'
+export { default } from './DisabledTaskPopupConnector'
