+++ packages/lib-classifier/src/components/Classifier/components/TaskArea/components/Tasks/index.js
@@ -1,1 +1,1 @@
-export { default } from './Tasks'
+export { default } from './TasksConnector'
