+++ packages/lib-classifier/src/components/Classifier/components/TaskArea/index.js
@@ -1,1 +1,1 @@
-export { default } from './TaskArea'
+export { default } from './TaskAreaConnector'
